+++ plugins/klogs/pkg/instance/helpers.go
@@ -1,9 +1,5 @@
 package instance
 
-import (
-	"sync"
-)
-
 // appendIfMissing appends a value to a slice, when this values doesn't exist in the slice already.
 func appendIfMissing(items []string, item string) []string {
 	for _, ele := range items {
@@ -25,19 +21,4 @@ func contains(items []string, item string) bool {
 	}
 
 	return false
-}
-
-// parallelize runs the given functions in parallel.
-func parallelize(functions ...func()) {
-	var waitGroup sync.WaitGroup
-	waitGroup.Add(len(functions))
-
-	defer waitGroup.Wait()
-
-	for _, function := range functions {
-		go func(copy func()) {
-			defer waitGroup.Done()
-			copy()
-		}(function)
-	}
 }